Save and restore DB comments
Summary: I've found this tool very useful (thank you for building it!),
but one issue I've run into is that when I save a snapshot and then
restore it later it will clear the comment which was present on the
database when it was snapshotted. We have a workflow in our dev
environment which depends on DB comments, which is why I've noticed this
behavior.

This updates the `save` and `restore` commands to also save DB comments.
When running `save` it will see if a comment exists on the DB being
snapshot, and if so set the same comment on the snapshot DB. When
running `restore` it will look for a comment on the snapshot DB being
restored, and set it on the newly created DB if it exists.

It might make sense to also show saved comments when running `list`, but
I didn't do that here.

I debated if this should be enabled via a flag people need to set to
opt-in to this behavior, but I didn't see a reason people who are using
DB comments wouldn't want this.

I tried using pq query parameters when specifying the comment, but it
seemed like that didn't work with pq (maybe because the `COMMENT ON...`
is an odd query), so instead I used `pq.QuoteLiteral` to escape the
comment.

Test Plan: tested running `save` and `restore` for a DB with a comment.
confirmed that the comment was restored correctly, even if it was
changed on the DB that was overwritten by the restore in the meantime.

tested running `save` and `restore` for a DB without any comment, and
confirmed both ran with no error, and no comment was saved on the DB or
the snapshot.